Issue #41
The receiver now correctly utilizies our message handler.
